Refactored scenarios
