+++ src/Battle.Tests/Scenarios/MultipleTurnsScenarioTest.cs
@@ -1,0 +1,19 @@
+﻿using Battle.Logic.Characters;
+using Battle.Logic.Encounters;
+using Battle.Logic.GameController;
+using Battle.Logic.Map;
+using Battle.Logic.Utility;
+using Battle.Tests.Characters;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System.Collections.Generic;
+using System.Numerics;
+
+namespace Battle.Tests.Scenarios
+{
+    [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
+    [TestClass]
+    [TestCategory("L0")]
+    public class MultipleTurnsScenarioTest
+    {
+    }
+}
